+++ src/MicroElements.Metadata.OpenXml/Reporting/Styling/StylingExtensions.cs
@@ -18,6 +18,78 @@ namespace MicroElements.Reporting.Styling
     public static class StylingExtensions
     {
         /// <summary>
+        /// Configures cell.
+        /// </summary>
+        /// <typeparam name="TMetadataProvider">MetadataProvider.</typeparam>
+        /// <param name="metadataProvider">Source metadata.</param>
+        /// <param name="configureCell">Cell customization action.</param>
+        /// <param name="combineMode">Combine mode. Default: AppendToEnd.</param>
+        /// <returns>The same metadata instance.</returns>
+        public static TMetadataProvider ConfigureCell<TMetadataProvider>(
+            this TMetadataProvider metadataProvider,
+            Action<CellContext> configureCell,
+            CombineMode combineMode = CombineMode.AppendToEnd)
+            where TMetadataProvider : IMetadataProvider
+        {
+            return metadataProvider.ConfigureMetadata<TMetadataProvider, ExcelCellMetadata>(
+                metadata => metadata.WithCombinedConfigure(ExcelCellMetadata.ConfigureCell, configureCell, combineMode));
+        }
+
+        /// <summary>
+        /// Configures column.
+        /// </summary>
+        /// <typeparam name="TMetadataProvider">MetadataProvider.</typeparam>
+        /// <param name="metadataProvider">Source metadata.</param>
+        /// <param name="configureColumn">Column customization action.</param>
+        /// <param name="combineMode">Combine mode. Default: AppendToEnd.</param>
+        /// <returns>The same metadata instance.</returns>
+        public static TMetadataProvider ConfigureColumn<TMetadataProvider>(
+            this TMetadataProvider metadataProvider,
+            Action<ColumnContext> configureColumn,
+            CombineMode combineMode = CombineMode.AppendToEnd)
+            where TMetadataProvider : IMetadataProvider
+        {
+            return metadataProvider.ConfigureMetadata<TMetadataProvider, ExcelColumnMetadata>(
+                metadata => metadata.WithCombinedConfigure(ExcelColumnMetadata.ConfigureColumn, configureColumn, combineMode));
+        }
+
+        /// <summary>
+        /// Configures column HeaderCell.
+        /// </summary>
+        /// <typeparam name="TMetadataProvider">MetadataProvider.</typeparam>
+        /// <param name="metadataProvider">Source metadata.</param>
+        /// <param name="configureHeaderCell">HeaderCell customization action.</param>
+        /// <param name="combineMode">Combine mode. Default: AppendToEnd.</param>
+        /// <returns>The same metadata instance.</returns>
+        public static TMetadataProvider ConfigureHeaderCell<TMetadataProvider>(
+            this TMetadataProvider metadataProvider,
+            Action<CellContext> configureHeaderCell,
+            CombineMode combineMode = CombineMode.AppendToEnd)
+            where TMetadataProvider : IMetadataProvider
+        {
+            return metadataProvider.ConfigureMetadata<TMetadataProvider, ExcelColumnMetadata>(
+                metadata => metadata.WithCombinedConfigure(ExcelColumnMetadata.ConfigureHeaderCell, configureHeaderCell, combineMode));
+        }
+
+        /// <summary>
+        /// Configures Row.
+        /// </summary>
+        /// <typeparam name="TMetadataProvider">MetadataProvider.</typeparam>
+        /// <param name="metadataProvider">Source metadata.</param>
+        /// <param name="configureRow">Row customization action.</param>
+        /// <param name="combineMode">Combine mode. Default: AppendToEnd.</param>
+        /// <returns>The same metadata instance.</returns>
+        public static TMetadataProvider ConfigureRow<TMetadataProvider>(
+            this TMetadataProvider metadataProvider,
+            Action<RowContext> configureRow,
+            CombineMode combineMode = CombineMode.AppendToEnd)
+            where TMetadataProvider : IMetadataProvider
+        {
+            return metadataProvider.ConfigureMetadata<TMetadataProvider, ExcelSheetMetadata>(
+                metadata => metadata.WithCombinedConfigure(ExcelSheetMetadata.ConfigureRow, configureRow, combineMode));
+        }
+
+        /// <summary>
         /// Gets format name for <see cref="ICellFormatProvider"/>.
         /// This name used for registering format in stylesheet.
         /// </summary>
@@ -34,12 +106,12 @@ namespace MicroElements.Reporting.Styling
         /// </summary>
         /// <param name="context">Cell context.</param>
         /// <param name="applyStyleName">Style name to apply.</param>
-        /// <param name="styleApply">Apply style.</param>
-        public static void ApplyStyleToCell(this CellContext context, string applyStyleName, StyleApply styleApply = StyleApply.Merge)
+        /// <param name="mergeMode">Apply style.</param>
+        public static void ApplyStyleToCell(this CellContext context, string applyStyleName, MergeMode mergeMode = MergeMode.Merge)
         {
             var documentContext = context.ColumnContext.SheetContext.DocumentContext;
 
-            context.Cell.StyleIndex = GetMergedStyleIndex(documentContext, context.Cell.StyleIndex, applyStyleName, styleApply);
+            context.Cell.StyleIndex = GetMergedStyleIndex(documentContext, context.Cell.StyleIndex, applyStyleName, mergeMode);
         }
 
         /// <summary>
@@ -48,12 +120,12 @@ namespace MicroElements.Reporting.Styling
         /// </summary>
         /// <param name="context">Cell context.</param>
         /// <param name="applyStyleName">Style name to apply.</param>
-        /// <param name="styleApply">Apply style.</param>
-        public static void ApplyStyleToColumn(this ColumnContext context, string applyStyleName, StyleApply styleApply = StyleApply.Merge)
+        /// <param name="mergeMode">Apply style.</param>
+        public static void ApplyStyleToColumn(this ColumnContext context, string applyStyleName, MergeMode mergeMode = MergeMode.Merge)
         {
             var documentContext = context.SheetContext.DocumentContext;
 
-            context.Column.Style = GetMergedStyleIndex(documentContext, context.Column.Style, applyStyleName, styleApply);
+            context.Column.Style = GetMergedStyleIndex(documentContext, context.Column.Style, applyStyleName, mergeMode);
         }
 
         /// <summary>
@@ -62,13 +134,13 @@ namespace MicroElements.Reporting.Styling
         /// <param name="documentContext">Document context.</param>
         /// <param name="currentStyleIndex">Current style index.</param>
         /// <param name="applyStyleName">New style name.</param>
-        /// <param name="styleApply">Apply style.</param>
+        /// <param name="mergeMode">Apply style.</param>
         /// <returns>Index of new style.</returns>
         public static uint GetMergedStyleIndex(
             this DocumentContext documentContext,
             UInt32Value? currentStyleIndex,
             string applyStyleName,
-            StyleApply styleApply = StyleApply.Merge)
+            MergeMode mergeMode = MergeMode.Merge)
         {
             int styleIndex = (int)(currentStyleIndex?.Value ?? 0);
 
@@ -77,9 +149,9 @@ namespace MicroElements.Reporting.Styling
             {
                 CellFormat currentStyle = documentContext.GetCellFormat(styleIndex);
                 string currentStyleName = currentStyle.GetStyleSheetName();
-                string combinedStyleName = styleApply == StyleApply.Merge
+                string combinedStyleName = mergeMode == MergeMode.Merge
                     ? (currentStyleName, applyStyleName).ToString()
-                    : styleApply == StyleApply.ReverseMerge
+                    : mergeMode == MergeMode.ReverseMerge
                         ? (applyStyleName, currentStyleName).ToString()
                         : applyStyleName;
 
@@ -92,9 +164,9 @@ namespace MicroElements.Reporting.Styling
                 else
                 {
                     CellFormat applyStyle = documentContext.GetCellFormat(applyStyleName);
-                    CellFormat combineStyle = styleApply == StyleApply.Merge
+                    CellFormat combineStyle = mergeMode == MergeMode.Merge
                         ? CombineStyles(currentStyle, applyStyle)
-                        : styleApply == StyleApply.ReverseMerge
+                        : mergeMode == MergeMode.ReverseMerge
                             ? CombineStyles(applyStyle, currentStyle)
                             : applyStyle;
 
@@ -166,31 +238,31 @@ namespace MicroElements.Reporting.Styling
         /// <typeparam name="TMetadataProvider">MetadataProvider.</typeparam>
         /// <param name="metadataProvider">Source metadata.</param>
         /// <param name="getCellStyle">Function. Input: CellValue, Result: StyleName.</param>
-        /// <param name="styleApply">StyleApply mode.</param>
-        /// <returns>The same metadata.</returns>
+        /// <param name="mergeMode">StyleApply mode. Default: Merge.</param>
+        /// <returns>The same metadata instance.</returns>
         public static TMetadataProvider ConfigureCellStyle<TMetadataProvider>(
             this TMetadataProvider metadataProvider,
             Func<string?, string?> getCellStyle,
-            StyleApply styleApply = StyleApply.Merge)
+            MergeMode mergeMode = MergeMode.Merge)
             where TMetadataProvider : IMetadataProvider
         {
             metadataProvider.AssertArgumentNotNull(nameof(metadataProvider));
             getCellStyle.AssertArgumentNotNull(nameof(getCellStyle));
 
-            if (styleApply == StyleApply.Set)
+            if (mergeMode == MergeMode.Set)
             {
                 metadataProvider.ConfigureMetadata<ExcelCellMetadata>(
-                    metadata => metadata.SetValue(ExcelCellMetadata.ConfigureCell, context => ConfigureCell(context, getCellStyle, styleApply)));
+                    metadata => metadata.SetValue(ExcelCellMetadata.ConfigureCell, context => ConfigureCell(context, getCellStyle, mergeMode)));
             }
             else
             {
                 metadataProvider.ConfigureMetadata<ExcelCellMetadata>(metadata =>
-                    metadata.WithCombinedConfigure(ExcelCellMetadata.ConfigureCell, context => ConfigureCell(context, getCellStyle, styleApply)));
+                    metadata.WithCombinedConfigure(ExcelCellMetadata.ConfigureCell, context => ConfigureCell(context, getCellStyle, mergeMode)));
             }
 
             return metadataProvider;
 
-            static void ConfigureCell(CellContext context, Func<string?, string?> getCellStyle, StyleApply styleApply)
+            static void ConfigureCell(CellContext context, Func<string?, string?> getCellStyle, MergeMode styleApply)
             {
                 string? cellValue = context.GetCellValue();
                 string? cellStyle = getCellStyle(cellValue);